ICU-12811 Refactor test-framework to not depend on core
Fixing the maven-build projects
