+++ icu4j/main/tests/core/src/com/ibm/icu/dev/test/ResourceModule.java
