+++ icu4j/main/tests/core/src/com/ibm/icu/dev/test/format/DateFormatTest.java
@@ -75,6 +75,7 @@ public class DateFormatTest extends TestFmwk {
      */
     @Test
     public void TestPatterns() {
+
         final String[][] EXPECTED = {
                 {DateFormat.YEAR, "y","en","y"},
 
@@ -2767,10 +2768,10 @@ public class DateFormatTest extends TestFmwk {
                 logln("time zone ex3 ok");
             }
             catch (Exception e) {
-                // hmmm... this shouldn't happen.  don't want to exit this
-                // fn with timezone improperly set, so just in case
-                TimeZone.setDefault(oldtz);
+                // hmmm... this shouldn't happen.
                 throw new IllegalStateException(e.getMessage());
+            } finally {
+                TimeZone.setDefault(oldtz);
             }
         }
