+++ icu4j/main/tests/core/src/com/ibm/icu/dev/test/serializable/SerializableWriter.java
@@ -71,11 +71,16 @@ public class SerializableWriter
 
         // Override default TimeZone, so serialized data always use
         // the consistent zone if not specified.
-        TimeZone.setDefault(TimeZone.getTimeZone("America/Los_Angeles"));
+        TimeZone savedZone = TimeZone.getDefault();
+        try {
+            TimeZone.setDefault(TimeZone.getTimeZone("America/Los_Angeles"));
 
-        SerializableWriter writer = new SerializableWriter(outDir);
+            SerializableWriter writer = new SerializableWriter(outDir);
 
-        writer.serialize();
+            writer.serialize();
+        } finally {
+            TimeZone.setDefault(savedZone);
+        }
     }
 
     public void serialize() throws IOException {
