+++ icu4j/main/tests/framework/src/com/ibm/icu/dev/test/AbstractTestLog.java
@@ -14,37 +14,7 @@ import java.util.GregorianCalendar;
 
 import org.junit.Assert;
 
-import com.ibm.icu.util.VersionInfo;
-
 public abstract class AbstractTestLog implements TestLog {
-    /**
-     * Returns true if ICU_Version < major.minor.
-     */
-    static public boolean isICUVersionBefore(int major, int minor) {
-        return isICUVersionBefore(major, minor, 0);
-    }
-
-    /**
-     * Returns true if ICU_Version < major.minor.milli.
-     */
-    static public boolean isICUVersionBefore(int major, int minor, int milli) {
-        return VersionInfo.ICU_VERSION.compareTo(VersionInfo.getInstance(major, minor, milli)) < 0;
-    }
-
-    /**
-     * Returns true if ICU_Version >= major.minor.
-     */
-    static public boolean isICUVersionAtLeast(int major, int minor) {
-        return isICUVersionAtLeast(major, minor, 0);
-    }
-
-    /**
-     * Returns true if ICU_Version >= major.minor.milli.
-     */
-    static public boolean isICUVersionAtLeast(int major, int minor, int milli) {
-        return !isICUVersionBefore(major, minor, milli);
-    }
-
     /**
      * Add a message.
      */
