+++ icu4j/main/tests/framework/src/com/ibm/icu/dev/test/TestUtil.java
@@ -15,10 +15,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.Locale;
-import java.util.MissingResourceException;
-
-import com.ibm.icu.impl.breakiter.LSTMBreakEngine;
-import com.ibm.icu.lang.UScript;
 
 public final class TestUtil {
     /**
@@ -282,29 +278,5 @@ public final class TestUtil {
             }
         }
         return ver;
-    }
-
-    private static boolean lstmDataIsBuilt() {
-        try {
-            LSTMBreakEngine.createData(UScript.THAI);
-            return true;
-        } catch (MissingResourceException e) {
-            // do nothing
-        }
-        try {
-            LSTMBreakEngine.createData(UScript.MYANMAR);
-            return true;
-        } catch (MissingResourceException e) {
-            // do nothing
-        }
-        return false;
-    }
-
-    public static boolean skipLSTMTest() {
-        return ! lstmDataIsBuilt();
-    }
-
-    public static boolean skipDictionaryTest() {
-        return lstmDataIsBuilt();
     }
 }
