+++ icu4j/main/tests/translit/.classpath
@@ -4,7 +4,11 @@
 	<classpathentry combineaccessrules="false" kind="src" path="/icu4j-translit"/>
 	<classpathentry combineaccessrules="false" kind="src" path="/icu4j-core"/>
 	<classpathentry combineaccessrules="false" kind="src" path="/icu4j-test-framework"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
+	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8">
+		<attributes>
+			<attribute name="module" value="true"/>
+		</attributes>
+	</classpathentry>
 	<classpathentry combineaccessrules="false" kind="src" path="/icu4j-langdata"/>
 	<classpathentry kind="lib" path="/external-libraries/hamcrest-core-1.3.jar"/>
 	<classpathentry kind="lib" path="/external-libraries/junit-4.12.jar" sourcepath="/external-libraries/junit-4.12-sources.jar">
@@ -17,5 +21,6 @@
 			<attribute name="javadoc_location" value="jar:platform:/resource/external-libraries/JUnitParams-1.0.5-javadoc.jar!/"/>
 		</attributes>
 	</classpathentry>
+	<classpathentry combineaccessrules="false" kind="src" path="/icu4j-core-tests"/>
 	<classpathentry kind="output" path="out/bin"/>
 </classpath>
