+++ icu4j/main/tests/translit/src/com/ibm/icu/dev/util/ImmutableEntry.java
